ref(discover) Extract table actions into separate component
Slim down the TableView by extracting the action buttons into a separate
component. I had to make some sacrifices to appease eslint proptype
checking.
